+++ src/Network/Network/Models/PSVirtualNetworkGateway.cs
@@ -88,6 +88,9 @@ namespace Microsoft.Azure.Commands.Network.Models
         public PSVirtualNetworkGatewayAutoscaleConfiguration AutoScaleConfiguration { get; set; }
 
         [Ps1Xml(Target = ViewControl.Table)]
+        public PSManagedServiceIdentity Identity { get; set; }
+
+        [Ps1Xml(Target = ViewControl.Table)]
         public PSVirtualNetworkGatewayMigrationStatus VirtualNetworkGatewayMigrationStatus { get; set; }
 
         [JsonIgnore]
@@ -148,6 +151,12 @@ namespace Microsoft.Azure.Commands.Network.Models
         public string AutoScaleConfigurationText
         {
             get { return JsonConvert.SerializeObject(AutoScaleConfiguration, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
+        }
+
+        [JsonIgnore]
+        public string IdentityText
+        {
+            get { return JsonConvert.SerializeObject(Identity, Formatting.Indented, new JsonSerializerSettings() { NullValueHandling = NullValueHandling.Ignore }); }
         }
 
         [JsonIgnore]